+++ tests/functional/test_experiment.py
@@ -138,3 +138,6 @@ class TestExperiment:
 
     def test_jupyter_notebooks(self, entity):
         assert sharedTests.jupyterNotebooks(entity)
+
+    def test_experiment_conditions(self, entity):
+        assert sharedTests.conditions(entity)