+++ tests/functional/test_jupyter_instance.py
@@ -18,21 +18,21 @@ class TestJupyterInstance:
         loadFixtures('Python\\\\JupyterInstance')
 
     def test_get(self, loggedUser):
-        guid = 'guid-test'
+        guid = 'd4b88c0b-37e8-4c84-8ce6-49a5661cd646'
         jupyterInstance = loggedUser.getJupyterInstance(guid)
         assert guid == jupyterInstance.guid
         assert jupyterInstance.started_at is None
         assert jupyterInstance.ended_at is None
 
     def test_start(self, loggedUser):
-        guid = 'guid-test'
+        guid = 'd4b88c0b-37e8-4c84-8ce6-49a5661cd646'
         jupyterInstance = loggedUser.getJupyterInstance(guid)
         assert jupyterInstance.started_at is None
         jupyterInstance.start()
         assert jupyterInstance.started_at is not None
 
     def test_end(self, loggedUser):
-        guid = 'guid-test'
+        guid = 'd4b88c0b-37e8-4c84-8ce6-49a5661cd646'
         jupyterInstance = loggedUser.getJupyterInstance(guid)
         assert jupyterInstance.ended_at is None
         jupyterInstance.end()